+++ plugin-dev/Source/Sentry/Private/Interface/SentryIdInterface.h
@@ -10,4 +10,5 @@ public:
 	virtual ~ISentryId() = default;
 
 	virtual FString ToString() const = 0;
+	virtual bool IsValid() const = 0;
 };