+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 ## Unreleased
 
+### Features
+
+- Add support for iOS Simulator builds on UE 5.5 and newer ([#1561](https://github.com/getsentry/sentry-unreal/pull/1561))
+- Add Windows Error Reporting support for the native backend ([#1568](https://github.com/getsentry/sentry-unreal/pull/1568))
+
+### Dependencies
+
+- Bump Java SDK from v8.54.0 to v8.55.0 ([#1560](https://github.com/getsentry/sentry-unreal/pull/1560))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8550)
+  - [diff](https://github.com/getsentry/sentry-java/compare/8.54.0...8.55.0)
+- Bump Android Gradle Plugin from v6.20.0 to v6.21.0 ([#1562](https://github.com/getsentry/sentry-unreal/pull/1562))
+  - [changelog](https://github.com/getsentry/sentry-android-gradle-plugin/blob/main/CHANGELOG.md#6210)
+  - [diff](https://github.com/getsentry/sentry-android-gradle-plugin/compare/6.20.0...6.21.0)
+- Bump Cocoa SDK from v9.26.1 to v9.27.0 ([#1563](https://github.com/getsentry/sentry-unreal/pull/1563))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#9270)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.26.1...9.27.0)
+- Bump Native SDK from v0.16.5 to v0.16.6 ([#1571](https://github.com/getsentry/sentry-unreal/pull/1571))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0166)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.16.5...0.16.6)
+
+## 1.22.0
+
+### Features
+
+- Add thread stackwalk mode setting for the native backend ([#1546](https://github.com/getsentry/sentry-unreal/pull/1546))
+
+### Fixes
+
+- Fix eventless attachments being uploaded on Mac/iOS ([#1558](https://github.com/getsentry/sentry-unreal/pull/1558), [#1559](https://github.com/getsentry/sentry-unreal/pull/1559))
+
 ### Dependencies
 
 - Bump Android Gradle Plugin from v6.19.0 to v6.20.0 ([#1553](https://github.com/getsentry/sentry-unreal/pull/1553))
@@ -13,6 +43,9 @@
 - Bump CLI from v3.6.2 to v3.7.0 ([#1554](https://github.com/getsentry/sentry-unreal/pull/1554))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#370)
   - [diff](https://github.com/getsentry/sentry-cli/compare/3.6.2...3.7.0)
+- Bump Native SDK from v0.16.4 to v0.16.5 ([#1557](https://github.com/getsentry/sentry-unreal/pull/1557))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0165)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.16.4...0.16.5)
 
 ## 1.21.0